Splitte OppgaveDao i tre: en for API, en for selve og en for utprøving
