+++ spesialist-api/src/main/kotlin/no/nav/helse/spesialist/api/oppgave/OppgaveApiDto.kt
@@ -8,12 +8,7 @@ import no.nav.helse.spesialist.api.vedtaksperiode.Periodetype
 import java.time.LocalDateTime
 import java.util.*
 
-data class PaginertOppgave(
-    val oppgave: OppgaveForOversiktsvisningDto,
-    val radnummer: Int,
-)
-
-data class OppgaveForOversiktsvisningDto (
+data class OppgaveForOversiktsvisningDto(
     val oppgavereferanse: String,
     val oppgavetype: String,
     val opprettet: LocalDateTime,
