+++ spesialist-selve/src/test/kotlin/no/nav/helse/e2e/OverstyringE2ETest.kt
@@ -43,7 +43,7 @@ internal class OverstyringE2ETest : AbstractE2ETest() {
 
     private fun List<OppgaveForOversiktsvisningDto>.ingenOppgaveMedId(id: String) = none { it.oppgavereferanse == id }
     private fun assertIngenOppgaver(id: String) {
-        oppgaveDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER).ingenOppgaveMedId(id)
+        oppgaveApiDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER).ingenOppgaveMedId(id)
     }
 
     @Test
@@ -75,7 +75,7 @@ internal class OverstyringE2ETest : AbstractE2ETest() {
         klargjørForGodkjenning(nyttGodkjenningsbehov)
 
         val oppgave =
-            oppgaveDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER).find { it.fødselsnummer == FØDSELSNUMMER }
+            oppgaveApiDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER).find { it.fødselsnummer == FØDSELSNUMMER }
         assertEquals(SAKSBEHANDLER_EPOST, oppgave!!.tildeling?.epost)
     }
 
@@ -116,7 +116,7 @@ internal class OverstyringE2ETest : AbstractE2ETest() {
 
         klargjørForGodkjenning(nyttGodkjenningsbehov)
 
-        val oppgave = requireNotNull(oppgaveDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER)
+        val oppgave = requireNotNull(oppgaveApiDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER)
             .find { it.fødselsnummer == FØDSELSNUMMER })
         assertEquals(SAKSBEHANDLER_EPOST, oppgave.tildeling?.epost)
     }
@@ -154,7 +154,7 @@ internal class OverstyringE2ETest : AbstractE2ETest() {
 
         klargjørForGodkjenning(nyttGodkjenningsbehov)
 
-        val oppgave = requireNotNull(oppgaveDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER)
+        val oppgave = requireNotNull(oppgaveApiDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER)
             .find { it.fødselsnummer == FØDSELSNUMMER })
         assertEquals(SAKSBEHANDLER_EPOST, oppgave.tildeling?.epost)
     }
@@ -246,7 +246,7 @@ internal class OverstyringE2ETest : AbstractE2ETest() {
 
         // TODO: bør ikke koble seg på daoer i E2E
         assertTrue(
-            oppgaveDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER).any { it.fødselsnummer == FØDSELSNUMMER })
+            oppgaveApiDao.finnOppgaver(SAKSBEHANDLERTILGANGER_UTEN_TILGANGER).any { it.fødselsnummer == FØDSELSNUMMER })
 
         val snapshot: Person = personQuery.person(FØDSELSNUMMER, null, dataFetchingEnvironment).data!!
